working/requirements.txt
tensorflow==2.12.0
numpy==1.26.4
pillow==10.2.0
flask==3.0.2
werkzeug==3.0.1
gunicorn==21.2.0
opencv-python==4.9.0.80
